+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 v 2.63
-  Added TextEditor toggle tipToggle highlight, mode label
+  Added TextEditor toggle tipToggle highlight, mode label,
+  Added underline button
 
 v 2.62
   Added TextEditor content limit,